+++ app/main/Constants/Locales.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace LaravelLang\Lang\Constants;
+
+final class Locales
+{
+    public const ENGLISH = 'en';
+}