+++ docs/changelog.md
@@ -13,6 +13,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
  * feature: [de] #1750 json file
  * feature: [nb] #1747 auth.php pagination.php passwords.php validation-inline.php validation.php json file
+ * feature: #1748 Split translation keys into Laravel packages
 
 ## [9.1.2] - 2021-05-14